GPS binary message refactoring work in progress
