+++ src/auto_config/auto_track.c
@@ -22,7 +22,7 @@
 #include "printk.h"
 #include "tracks.h"
 
-const Track* findClosestTrack(const Tracks *tracks, const GeoPoint location) {
+static const Track* findClosestTrack(const Tracks *tracks, const GeoPoint *location) {
     float dist = MAX_DIST_FROM_SF;
     const Track *best = NULL;
 
@@ -31,7 +31,7 @@ const Track* findClosestTrack(const Tracks *tracks, const GeoPoint location) {
 
         // XXX: inaccurate but fast.  Good enough for now.
         GeoPoint startPoint = getStartPoint(track);
-        float track_distance = distPythag(&startPoint, &location);
+        float track_distance = distPythag(&startPoint, location);
 
         if (track_distance >= dist)
             continue;
@@ -44,7 +44,7 @@ const Track* findClosestTrack(const Tracks *tracks, const GeoPoint location) {
     return best;
 }
 
-const Track* auto_configure_track(Track *defaultCfg, GeoPoint gp) {
+const Track* auto_configure_track(Track *defaultCfg, GeoPoint *gp) {
     pr_info("Configuring start/finish...\r\n");
 
     if (isStartPointValid(defaultCfg) && isFinishPointValid(defaultCfg)) {
