+++ src/gps/geoCircle.c
@@ -31,8 +31,8 @@ struct GeoCircle gc_createGeoCircle(const GeoPoint gp, const float r) {
    return gc;
 }
 
-bool gc_isPointInGeoCircle(const GeoPoint point, const struct GeoCircle gc) {
-   float dist = distPythag(&point, &(gc.point));
+bool gc_isPointInGeoCircle(const GeoPoint * point, const struct GeoCircle gc) {
+   float dist = distPythag(point, &(gc.point));
    return  dist <= gc.radius;
 }
