+++ src/gps/gps.c
@@ -14,7 +14,6 @@
 #include "printk.h"
 #include "tracks.h"
 
-
 #include <stdint.h>
 
 #define GPS_LOCK_FLASH_COUNT 5
@@ -36,15 +35,6 @@ static int g_flashCount;
 static float g_prevLatitude;
 static float g_prevLongitude;
 
-static float g_latitude;
-static float g_longitude;
-
-static float g_speed;
-
-static enum GpsSignalQuality g_gpsQuality;
-
-static int g_satellitesUsedForPosition;
-
 static int g_atStartFinish;
 static int g_prevAtStartFinish;
 static tiny_millis_t g_lastStartFinishTimestamp;
@@ -62,13 +52,6 @@ static tiny_millis_t g_lastSectorTime;
 static int g_lapCount;
 static float g_distance;
 
-/**
- * Date and time this GPS fix was taken.
- */
-static DateTime g_dtFirstFix;
-static DateTime g_dtLastFix;
-static millis_t g_utcMillisAtSample;
-static tiny_millis_t g_uptimeAtSample;
 
 static float degreesToMeters(float degrees) {
    // There are 110574.27 meters per degree of latitude at the equator.
@@ -101,8 +84,9 @@ static void updateMillisSinceEpoch(DateTime fixDateTime) {
 
 millis_t getMillisSinceEpoch() {
    // If we have no GPS data, return 0 to indicate that.
-   if (isGpsDataCold())
+   if (isGpsDataCold()){
       return 0;
+   }
 
    const tiny_millis_t deltaSinceSample = getUptime() - g_uptimeAtSample;
    return g_utcMillisAtSample + deltaSinceSample;
@@ -118,288 +102,6 @@ static void updateUptimeAtSample() {
 
 tiny_millis_t getUptimeAtSample() {
    return g_uptimeAtSample;
-}
-
-/**
- * Performs a full update of the g_dtLastFix value.  Also update the g_dtFirstFix value if it hasn't
- * already been set along with updating the Milliseconds as well
- * @param fixDateTime The DateTime of the GPS fix.
- */
-void updateFullDateTime(DateTime fixDateTime) {
-   updateUptimeAtSample();
-   updateMillisSinceEpoch(fixDateTime);
-
-   g_dtLastFix = fixDateTime;
-
-   if (g_dtFirstFix.year == 0)
-      g_dtFirstFix = fixDateTime;
-}
-
-/**
- * Like atoi, but is non-destructive to the string passed in and provides an offset and length
- * functionality.  Max Len is 3.
- * @param str The start of the String to parse.
- * @param offset How far in to start reading the string.
- * @param len The number of characters to read.
- */
-int atoiOffsetLenSafe(const char *str, size_t offset, size_t len) {
-   char buff[4] = { 0 };
-
-   // Bounds check.  Don't want any bleeding hearts in here...
-   if (len > (sizeof(buff) - 1))
-         len = sizeof(buff) - 1;
-
-   memcpy(buff, str + offset, len);
-   return modp_atoi(buff);
-}
-
-static float parseLatitude(char *data){
-   unsigned int len = strlen(data);
-
-   if (len <= 0 && len > LATITUDE_DATA_LEN)
-      return 0.0;
-
-   //Raw GPS Format is ddmm.mmmmmm
-   int degrees = atoiOffsetLenSafe(data, 0, 2);
-   float minutes = modp_atof(data + 2);
-
-   return degrees + (minutes / 60);
-}
-
-static float parseLongitude(char *data){
-   unsigned int len = strlen(data);
-
-   if (len <= 0 && len > LONGITUDE_DATA_LEN)
-      return 0.0;
-
-   //Raw GPS Format is dddmm.mmmmmm
-   int degrees = atoiOffsetLenSafe(data, 0, 3);
-   float minutes = modp_atof(data + 3);
-
-   return degrees + (minutes / 60);
-}
-
-//Parse Global Positioning System Fix Data.
-static void parseGGA(char *data) {
-   char * delim = strchr(data, ',');
-   int param = 0;
-   int keepParsing = 1;
-
-   while (delim != NULL && keepParsing) {
-      *delim = '\0';
-      switch (param) {
-         /*
-      case 1: {
-         unsigned int len = strlen(data);
-         if (len > 0 && len <= LATITUDE_DATA_LEN) {
-            //Raw GPS Format is ddmm.mmmmmm
-            char degreesStr[3] = { 0 };
-            memcpy(degreesStr, data, 2);
-            degreesStr[2] = 0;
-            float minutes = modp_atof(data + 2);
-            minutes = minutes / 60.0;
-            latitude = modp_atoi(degreesStr) + minutes;
-         } else {
-            latitude = 0;
-            //TODO log error
-         }
-      }
-         break;
-      case 2: {
-         if (data[0] == 'S') {
-            latitude = -latitude;
-         }
-      }
-         break;
-      case 3: {
-         unsigned int len = strlen(data);
-         if (len > 0 && len <= LONGITUDE_DATA_LEN) {
-            //Raw GPS Format is dddmm.mmmmmm
-            char degreesStr[4] = { 0 };
-            memcpy(degreesStr, data, 3);
-            degreesStr[3] = 0;
-            float minutes = modp_atof(data + 3);
-            minutes = minutes / 60.0;
-            longitude = modp_atoi(degreesStr) + minutes;
-         } else {
-            longitude = 0;
-            //TODO log error
-         }
-      }
-         break;
-      case 4: {
-         if (data[0] == 'W') {
-            longitude = -longitude;
-         }
-      }
-         break;
-      case 5:
-         g_gpsQuality = (enum GpsSignalQuality) modp_atoi(data);
-         break;
-      case 6:
-         g_satellitesUsedForPosition = modp_atoi(data);
-         keepParsing = 0;
-         break;
-         */ // Kept until Full sanity checks pass.
-
-      case 5:
-         // FIXME: Better suport fo GPS quality here?
-         g_gpsQuality = modp_atoi(data) > 0 ? GPS_QUALITY_FIX : GPS_QUALITY_NO_FIX;
-         break;
-
-      case 6:
-         g_satellitesUsedForPosition = modp_atoi(data);
-         keepParsing = 0;
-         break;
-
-      }
-
-      param++;
-      data = delim + 1;
-      delim = strchr(data, ',');
-   }
-}
-
-void updatePosition(float latitude, float longitude) {
-   g_prevLatitude = g_latitude;
-   g_prevLongitude = g_longitude;
-   g_longitude = longitude;
-   g_latitude = latitude;
-}
-
-//Parse GNSS DOP and Active Satellites
-static void parseGSA(char *data) {}
-
-//Parse Course Over Ground and Ground Speed
-static void parseVTG(char *data) {
-
-   char * delim = strchr(data, ',');
-   int param = 0;
-
-   int keepParsing = 1;
-   while (delim != NULL && keepParsing) {
-      *delim = '\0';
-      switch (param) {
-      case 6: //Speed over ground
-      {
-         if (strlen(data) >= 1) {
-            setGPSSpeed(modp_atof(data)); //convert to MPH
-         }
-         keepParsing = 0;
-      }
-         break;
-      default:
-         break;
-      }
-      param++;
-      data = delim + 1;
-      delim = strchr(data, ',');
-   }
-}
-
-//Parse Geographic Position - Latitude / Longitude
-static void parseGLL(char *data) {
-
-}
-
-//Parse Time & Date
-static void parseZDA(char *data) {
-   /*
-    $GPZDA
-
-    Date & Time
-
-    UTC, day, month, year, and local time zone.
-
-    $--ZDA,hhmmss.ss,xx,xx,xxxx,xx,xx
-    hhmmss.ss = UTC
-    xx = Day, 01 to 31
-    xx = Month, 01 to 12
-    xxxx = Year
-    xx = Local zone description, 00 to +/- 13 hours
-    xx = Local zone minutes description (same sign as hours)
-    */
-   return;
-}
-
-//Parse GNSS Satellites in View
-static void parseGSV(char *data) {
-
-}
-
-//Parse Recommended Minimum Navigation Information
-static void parseRMC(char *data) {
-   /*
-    * $GPRMC,053740.000,A,2503.6319,N,12136.0099,E,2.69,79.65,100106,,,A*53
-    * Message ID $GPRMC RMC protocol header
-    * UTC Time 053740.000 hhmmss.sss
-    * Status A A=data valid or V=data not valid
-    * Latitude 2503.6319 ddmm.mmmm
-    * N/S Indicator N N=north or S=south
-    * Longitude 12136.0099 dddmm.mmmm
-    * E/W Indicator E E=east or W=west
-    * Speed over ground 2.69 knots True
-    * Course over ground 79.65 degrees
-    * Date 100106 Ddmmyy
-    * Magnetic variation degrees Not shown
-    * Variation sense E=east or W=west (Not shown)
-    * Mode A
-    * A=autonomous, D=DGPS, E=DR, N=Data not valid,
-    * R=Coarse Position, S=Simulator
-    * Checksum *53
-    * <CR> <LF> End of message termination
-    */
-
-   char *delim = strchr(data, ',');
-   int param = 0;
-   DateTime dt = { 0 };
-   float latitude = 0.0;
-   float longitude = 0.0;
-
-   while (delim) {
-      *delim = '\0';
-      switch (param) {
-      case 0: //UTC Time (HHMMSS.SSS)
-         dt.hour = (int8_t) atoiOffsetLenSafe(data, 0, 2);
-         dt.minute = (int8_t) atoiOffsetLenSafe(data, 2, 2);
-         dt.second = (int8_t) atoiOffsetLenSafe(data, 4, 2);
-         dt.millisecond = (int16_t) atoiOffsetLenSafe(data, 7, 3);
-         break;
-
-      case 2:
-    	 latitude = parseLatitude(data);
-         break;
-
-      case 3:
-         if (data[0] == 'S') latitude = -latitude;
-         break;
-
-      case 4:
-    	 longitude = parseLongitude(data);
-         break;
-
-      case 5:
-         if (data[0] == 'W') longitude = -longitude;
-         break;
-
-      case 6: //Speed over ground
-         if (strlen(data) >= 1) setGPSSpeed(modp_atof(data) * KNOTS_TO_KPH);
-         break;
-
-      case 8: //Date (DDMMYY)
-         dt.day = (int8_t) atoiOffsetLenSafe(data, 0, 2);
-         dt.month = (int8_t) atoiOffsetLenSafe(data, 2, 2);
-         dt.year = (int16_t) atoiOffsetLenSafe(data, 4, 2) + 2000;
-         break;
-      }
-
-      ++param;
-      data = ++delim;
-      delim = strchr(delim, ',');
-   }
-
-   updateFullDateTime(dt);
-   updatePosition(latitude, longitude);
 }
 
 void resetGpsDistance() {
@@ -494,19 +196,6 @@ DateTime getLastFixDateTime() {
    return g_dtLastFix;
 }
 
-static struct GpsSample getGpsSample() {
-   struct GpsSample s;
-
-   s.quality = getGPSQuality();
-   s.point.latitude = getLatitude();
-   s.point.longitude = getLongitude();
-   s.time = getMillisSinceEpoch();
-   s.firstFixMillis = getMillisSinceFirstFix();
-   s.speed = getGPSSpeed();
-
-   return s;
-}
-
 /**
  * @return True if we have crossed the start line at least once, false otherwise.
  */
@@ -514,23 +203,21 @@ static bool isStartCrossedYet() {
    return g_lastStartFinishTimestamp != 0;
 }
 
-static float calcDistancesSinceLastSample() {
+static float calcDistancesSinceLastSample(GpsSamp *gpsSample) {
    const GeoPoint prev = {g_prevLatitude, g_prevLongitude};
-   const GeoPoint curr = {g_latitude, g_longitude};
 
-   if (!isValidPoint(&prev) || !isValidPoint(&curr))
+   if (!isValidPoint(&prev) || !isValidPoint(&gpsSample->point)){
       return 0.0;
-
+   }
    // Return distance in KM
-   return distPythag(&prev, &curr) / 1000;
+   return distPythag(&prev, &gpsSample->point) / 1000;
 }
 
-static int processStartFinish(const Track *track, const float targetRadius) {
-   const struct GpsSample gpss = getGpsSample();
+static int processStartFinish(const GpsSamp *gpsSample, const Track *track, const float targetRadius) {
 
    // First time crossing start finish.  Handle this in a special way.
    if (!isStartCrossedYet()) {
-      lc_supplyGpsSample(gpss);
+      lc_supplyGpsSample(gpsSample);
 
       if (lc_hasLaunched()) {
          g_lastStartFinishTimestamp = lc_getLaunchTime();
@@ -556,7 +243,7 @@ static int processStartFinish(const Track *track, const float targetRadius) {
     * FIXME: Should have logic that checks that we left the start/finish circle
     * for some time.
     */
-   g_atStartFinish = gc_isPointInGeoCircle(gpss.point, sfCircle);
+   g_atStartFinish = gc_isPointInGeoCircle(&(gpsSample->point), sfCircle);
 
    if (!g_atStartFinish || g_prevAtStartFinish != 0 ||
        elapsed <= START_FINISH_TIME_THRESHOLD) {
@@ -619,11 +306,6 @@ void initGPS() {
    g_flashCount = 0;
    g_prevLatitude = 0.0;
    g_prevLongitude = 0.0;
-   g_latitude = 0.0;
-   g_longitude = 0.0;
-   g_gpsQuality = GPS_QUALITY_NO_FIX;
-   g_satellitesUsedForPosition = 0;
-   g_speed = 0.0;
    g_lastLapTime = 0;
    g_lastSectorTime = 0;
    g_atStartFinish = 0;
@@ -641,12 +323,12 @@ void initGPS() {
    g_uptimeAtSample = 0;
 }
 
-static void flashGpsStatusLed() {
+static void flashGpsStatusLed(enum GpsSignalQuality gpsQuality) {
    if (g_flashCount == 0)
       LED_disable(1);
    g_flashCount++;
 
-   int targetFlashCount = isGpsSignalUsable(g_gpsQuality) ?
+   int targetFlashCount = isGpsSignalUsable(gpsQuality) ?
       GPS_LOCK_FLASH_COUNT : GPS_NOFIX_FLASH_COUNT;
 
    if (g_flashCount >= targetFlashCount) {
@@ -672,16 +354,16 @@ static int isSectorTrackingEnabled(const Track *track) {
             isValidPoint(&p0) && isStartFinishEnabled(track);
 }
 
-void onLocationUpdated() {
+static void onLocationUpdated(GpsSamp *gpsSample) {
    static int sectorEnabled = 0;
    static int startFinishEnabled = 0;
 
    // If no GPS lock, no point in doing any of this.
-   if (!isGpsSignalUsable(g_gpsQuality))
+   if (!isGpsSignalUsable(gpsSample->quality))
       return;
 
    LoggerConfig *config = getWorkingLoggerConfig();
-   const GeoPoint gp = getGeoPoint();
+   const GeoPoint *gp = gpsSample->point;
 
    // FIXME: Improve on this.  Doesn't need calculation every time.
    const float targetRadius = degreesToMeters(config->TrackConfigs.radius);
@@ -696,9 +378,8 @@ void onLocationUpdated() {
       g_configured = 1;
    }
 
-   float dist = calcDistancesSinceLastSample();
+   float dist = calcDistancesSinceLastSample(gpsSample);
    g_distance += dist;
-
 
    if (startFinishEnabled) {
       // Seconds since first fix is good until we alter the code to use millis directly
@@ -715,7 +396,7 @@ void onLocationUpdated() {
          if (getLapCount() == 0) {
             const GeoPoint sp = getStartPoint(g_activeTrack);
             // Distance is in KM
-            g_distance = distPythag(&sp, &gp) / 1000;
+            g_distance = distPythag(&sp, gp) / 1000;
 
             startFinishCrossed(sp, g_lastStartFinishTimestamp);
             addGpsSample(gp, millisSinceFirstFix);
@@ -729,6 +410,9 @@ void onLocationUpdated() {
       if (sectorEnabled)
          processSector(g_activeTrack, targetRadius);
    }
+
+   g_prevLatitude = gpsSample->point.latitude;
+   g_prevLongitude = gpsSample->point.longitude;
 
 }
 
@@ -753,36 +437,10 @@ int checksumValid(const char *gpsData, size_t len) {
    return valid;
 }
 
-void processGPSData(char *gpsData, size_t len) {
-   if (len <= 4 || !checksumValid(gpsData, len) || strstr(gpsData, "$GP") != gpsData) {
-      pr_trace("GPS: corrupt frame ");
-      pr_trace(gpsData);
-      pr_trace("\r\n");
-      return;
-   }
-   size_t positionUpdated = 0;
-
-   // Advance the pointer 3 spaces since we know it begins with "$GP"
-   gpsData += 3;
-   if (strstr(gpsData, "GGA,")) {
-      parseGGA(gpsData + 4);
-   } else if (strstr(gpsData, "VTG,")) { //Course Over Ground and Ground Speed
-      parseVTG(gpsData + 4);
-   } else if (strstr(gpsData, "GSA,")) { //GPS Fix gpsData
-      parseGSA(gpsData + 4);
-   } else if (strstr(gpsData, "GSV,")) { //Satellites in view
-      parseGSV(gpsData + 4);
-   } else if (strstr(gpsData, "RMC,")) { //Recommended Minimum Specific GNSS Data
-      parseRMC(gpsData + 4);
-      positionUpdated = 1;
-   } else if (strstr(gpsData, "GLL,")) { //Geographic Position - Latitude/Longitude
-      parseGLL(gpsData + 4);
-   } else if (strstr(gpsData, "ZDA,")) { //Time & Date
-      parseZDA(gpsData + 4);
-   }
-
-   if (positionUpdated && !isGpsDataCold()) {
-      onLocationUpdated();
-      flashGpsStatusLed();
-   }
+void processGPSUpdate(GpsSamp *gpsSample){
+	if (!isGpsDataCold()){
+		  onLocationUpdated(gpsSample);
+		  flashGpsStatusLed(gpsSample->quality);
+	}
 }
+
